+++ tests/test_command_installation.py
@@ -16,6 +16,7 @@ COMMANDS = (
     "open-mmi-launcher",
     "open-mmi-status",
     "open-mmi-update-coordinator",
+    "open-mmi-update-installer",
 )
 
 
@@ -127,7 +128,7 @@ class CommandInstallationTests(unittest.TestCase):
 
                 test -f \"$COMMAND_LINK_DIR/open-mmi-launcher\"
                 test ! -L \"$COMMAND_LINK_DIR/open-mmi-launcher\"
-                for command in open-mmi-canbusd open-mmi-config open-mmi-dashboard open-mmi-status open-mmi-update-coordinator; do
+                for command in open-mmi-canbusd open-mmi-config open-mmi-dashboard open-mmi-status open-mmi-update-coordinator open-mmi-update-installer; do
                     test ! -e \"$COMMAND_LINK_DIR/$command\"
                     test ! -L \"$COMMAND_LINK_DIR/$command\"
                 done